+++ server/src/update_batcher.rs
@@ -718,6 +718,7 @@ mod tests {
             rich_presence: None,
             combat_achievements: None,
             slayer_task: None,
+            slayer_task_events: None,
             portrait_last_update: None,
             last_updated: None,
             events: None,